[백준]4779/1시간+30분/정답(반복횟수+Syso 시간초과 해결)
